requirements.txt
# pip install -r requirements.txt
couchdb
tweepy
listener
numpy
websocket
#twint
git+https://github.com/twintproject/twint.git@origin/master#egg=twint

